requirements.txt
dash[diskcache]==2.18.2
dash-bootstrap-components
pandas
numpy
plotly
openai
openpyxl
Flask-Caching
dash-bootstrap-templates
flask
gunicorn
psycopg2-binary
python-dotenv
flask_session
bcrypt